Enable Android compilation.
This is different from the capstone_android_port branch. This one
branches out from the "next" branch of Capstone. I tried to replay the
commits of the "capstone_android_port" branch on top of "next" but there
were conflicts that didn't appear easy to resolve, so I started out a
new branch instead.